/W3 found 858 warnings where /W1 found three, and four were real
Raising MSVC's warning level was worth it immediately.

  850 C4996  deprecated CRT function
    8 C4267  conversion from size_t to int, possible loss of data

The eight are four sites counted twice, and they are mine. SO() expands
to CchArray(sz) minus a pointer difference, which stays size_t, so every
SO() passing an int parameter narrows -- at general.cpp:1399 and 1403 and
xcharts0.cpp:486 and 490, all four feeding PchFormatExpression() and
PchFormatString(). S() has carried an explicit (int) since it was written;
SO() never did, and no GCC build mentions it. Now they match.

The 850 are MSVC deprecating fopen, getenv, strcpy and sscanf in favour
of _s variants that exist only on Microsoft's compiler. That is a
portability opinion rather than a defect: the code is correct standard C,
and adopting _s would make it non-portable to the four other compilers
that build this tree. _CRT_SECURE_NO_WARNINGS, with the reasoning written
where the flag is, because a silenced warning with no explanation is
indistinguishable from one nobody noticed.

Worth stating plainly what /W1 cost: for as long as this job existed, "the
MSVC build is clean" meant three warnings at the quietest setting the
compiler has. Four narrowing conversions in this codebase's own formatting
code were sitting under it the whole time.

Five builds, no new warnings, suite 3812/0.

+++ astrolog.h
@@ -461,7 +461,12 @@
 // named. That one sizeof cannot catch; this one it can, so it should.
 template <class T, size_t N> inline size_t CchArray(T (&)[N]) { return N; }
 #define S(sz) (sz), (int)CchArray(sz)
-#define SO(pch, sz) (pch), (CchArray(sz) - ((pch) - (sz)))
+// (int) on SO too, for the same reason S() has it: CchArray() returns
+// size_t and the subtraction stays size_t, so every SO() call passing an
+// int parameter is a narrowing conversion. GCC says nothing; MSVC says
+// C4267 "possible loss of data" -- but only at /W3, which is why four
+// sites sat unmentioned while the job ran at /W1.
+#define SO(pch, sz) (pch), (int)(CchArray(sz) - ((pch) - (sz)))
 #else
 #define S(sz) (sz), (int)sizeof(sz)
 #define SO(pch, sz) (pch), (sizeof(sz) - ((pch) - (sz)))
